requirements.txt
# Core UI
streamlit>=1.32.0

# LLM + Agent framework
langchain>=0.1.20
langgraph>=0.0.40
langchain-openai>=0.1.6

# Vector store
faiss-cpu>=1.8.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.3

# Utilities
python-dotenv>=1.0.1
pydantic>=2.6.0

# Database (builtin, but listed for clarity)
sqlite-utils>=3.36

# Google Calendar Integration
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.116.0
langchain-google-genai>=1.0.0
